+++ GridMvc/Pagination/IGridPager.cs
@@ -25,6 +25,11 @@ namespace GridMvc.Pagination
         int? ItemsCountOverwrite { get; set; }
 
         /// <summary>
+        ///     Disables default paging if set to true
+        /// </summary>
+        bool? CustomPaging { get; set; }
+
+        /// <summary>
         ///     Method invokes before pager render
         /// </summary>
         void Initialize<T>(IQueryable<T> items);